+++ docs/source/releasenotes.rst
@@ -1,6 +1,10 @@
 Release notes
 =============
 
+<next release>
+--------------
+- Images: Don't template match same region multiple times
+
 0.7.2
 -----
 - Browser: Store webdrivers in temporary directory